sequence value generator

+++ generator/ValueGenerator.go
@@ -1,6 +1,6 @@
 package generator
 
 type ValueGenerator interface {
-	ReserveRange(sequence string, count int)
+	ReserveRange(sequence string, count uint64)
 	GetNextValue(sequence string) string
 }
