+++ service/Generator.go
@@ -27,7 +27,7 @@ func (generator *Generator) Generate(r *http.Request, args *GenerateCommandArgum
 	}
 
 	ids := make([]string, 0)
-	generator.valueGenerator.ReserveRange(sequence, args.Count)
+	generator.valueGenerator.ReserveRange(sequence, uint64(args.Count))
 
 	for i := 0; i < args.Count; i++ {
 		ids = append(ids, generator.valueGenerator.GetNextValue(sequence))
